+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/FlutterAdManagerInterstitialAd.java
@@ -67,6 +67,7 @@ class FlutterAdManagerInterstitialAd extends FlutterAd.FlutterOverlayAd {
   void onAdLoaded(AdManagerInterstitialAd ad) {
     this.ad = ad;
     ad.setAppEventListener(new DelegatingAdManagerInterstitialAdCallbacks(this));
+      ad.setOnPaidEventListener(new FlutterPaidEventListener(manager, this));
     manager.onAdLoaded(adId, ad.getResponseInfo());
   }