+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/FlutterInterstitialAd.java
@@ -57,6 +57,7 @@ class FlutterInterstitialAd extends FlutterAd.FlutterOverlayAd {
 
   void onAdLoaded(InterstitialAd ad) {
     this.ad = ad;
+    interstitialAd.setOnPaidEventListener(new FlutterPaidEventListener(manager, this));
     manager.onAdLoaded(adId, ad.getResponseInfo());
   }