+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/FlutterRewardedAd.java
@@ -124,6 +124,7 @@ class FlutterRewardedAd extends FlutterAd.FlutterOverlayAd {
       rewardedAd.setServerSideVerificationOptions(
           serverSideVerificationOptions.asServerSideVerificationOptions());
     }
+    rewardedAd.setOnPaidEventListener(new FlutterPaidEventListener(manager, this));
     manager.onAdLoaded(adId, rewardedAd.getResponseInfo());
   }