+++ packages/google_mobile_ads/ios/Tests/FLTBannerAdTest.m
@@ -62,6 +62,24 @@
   [bannerAd.bannerView.delegate bannerViewDidRecordImpression:bannerMock];
   OCMVerify([mockManager onBannerImpression:[OCMArg isEqual:bannerAd]]);
 
+  // Mock callback of paid event handler.
+  GADAdValue *adValue = OCMClassMock([GADAdValue class]);
+  OCMStub([adValue value]).andReturn(NSDecimalNumber.one);
+  OCMStub([adValue precision]).andReturn(GADAdValuePrecisionEstimated);
+  OCMStub([adValue currencyCode]).andReturn(@"currencyCode");
+
+  bannerAd.bannerView.paidEventHandler(adValue);
+  OCMVerify([mockManager onPaidEvent:[OCMArg isEqual:bannerAd]
+                               value:[OCMArg checkWithBlock:^BOOL(id obj) {
+                                 FLTAdValue *adValue = obj;
+                                 XCTAssertEqualObjects(
+                                     adValue.valueMicros,
+                                     [[NSDecimalNumber alloc] initWithInt:1000000]);
+                                 XCTAssertEqual(adValue.precision, GADAdValuePrecisionEstimated);
+                                 XCTAssertEqualObjects(adValue.currencyCode, @"currencyCode");
+                                 return TRUE;
+                               }]]);
+
   NSString *domain = @"domain";
   NSDictionary *userInfo = @{NSLocalizedDescriptionKey : @"description"};
   NSError *error = [NSError errorWithDomain:domain code:1 userInfo:userInfo];